main/studymate-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymongo
google-cloud-storage
google-cloud-texttospeech
google-generativeai
pdfplumber
firebase-admin